+++ network-mux/test/Test/Mux.hs
@@ -11,6 +11,9 @@
 {-# LANGUAGE TypeFamilies               #-}
 
 {-# OPTIONS_GHC -Wno-orphans            #-}
+#if __GLASGOW_HASKELL__ >= 908
+{-# OPTIONS_GHC -Wno-x-partial          #-}
+#endif
 
 module Test.Mux (tests) where