+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Testnet.hs
@@ -9,6 +9,9 @@
 #if defined(mingw32_HOST_OS)
 {-# OPTIONS_GHC -Wno-unused-top-binds #-}
 #endif
+#if __GLASGOW_HASKELL__ >= 908
+{-# OPTIONS_GHC -Wno-x-partial #-}
+#endif
 
 module Test.Ouroboros.Network.Testnet (tests) where